Refactored evaluables and the color editor
With the upgraded indexable colorVec class I was passing around color
vecs much more often, what I failed to remeber was that this was
creating a mess of references. In one case I was using the color pulled
from EvaluateColor and using the reference in an editor, I decided this
was the wrong way to go about using evaluate color and all evaluate
color methods now return a copy of the color. Evaluate color shouldn't
be used as a way to alter the sate of an evaluable, that creates a mess.